read game data from JSON files

+++ core/src/main/scala/IntergalacticInterlopers.scala
@@ -9,8 +9,8 @@ import com.badlogic.gdx.graphics.glutils.ShapeRenderer._
 
 class IntergalacticInterlopers extends Game {
   class DemoScreen extends Screen {
-    val camera = new OrthographicCamera(Gdx.graphics.getWidth, Gdx.graphics.getHeight)
-    camera.position.set(Gdx.graphics.getWidth/2, Gdx.graphics.getHeight/2, 0)
+    val camera = new OrthographicCamera(Gdx.graphics.getWidth.toFloat, Gdx.graphics.getHeight.toFloat)
+    camera.position.set(Gdx.graphics.getWidth/2f, Gdx.graphics.getHeight/2f, 0)
     camera.update()
     val mainBatch = new SpriteBatch
     val shapeRenderer = new ShapeRenderer
